Removed abstract Basic Player class
<organisation
